APU: fix wave channel

+++ Devices/Audio/FrequencyCounter.cpp
@@ -2,13 +2,13 @@
 #include "Devices/Audio/FrequencyCounter.hpp"
 #include "Util/LogUtil.hpp"
 
-void FrequencyCounter::setPeriod(int period) {
+void FrequencyCounter::setPeriod(float period) {
     this->period = period;
     this->value = period;
     this->outputClockTicks = 0;
 }
 
-void FrequencyCounter::updatePeriod(int period) {
+void FrequencyCounter::updatePeriod(float period) {
     this->period = period;
 }
 
@@ -17,7 +17,9 @@ long FrequencyCounter::getOutputClockTicks() {
 }
 
 bool FrequencyCounter::update() {
-    int newValue = value - MASTER_CLOCK_TICKS_PER_SAMPLE;
+    
+    
+    float newValue = value - MASTER_CLOCK_TICKS_PER_SAMPLE;
     if (newValue <= 0) {
         value = period + newValue;
         outputClockTicks++;
